Astro-Insight-0913v3/requirements.txt
# 天体分析系统 - 综合Python依赖包
# 整合了主系统和MCP_retrieval检索模块的所有依赖

# ===========================================
# 核心框架和基础库
# ===========================================

# Web框架
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# 天文学专业库
astropy==5.3.4
astroquery==0.4.6
specutils==1.11.0
numpy==1.24.4
matplotlib==3.7.2
pandas==2.0.3
scipy==1.11.4

# 机器学习
scikit-learn==1.3.2
joblib==1.3.2

# 深度学习框架（multimark模块）
tensorflow>=2.13.0
keras>=2.13.0

# 数据可视化（可选，用于高级可视化）
# seaborn>=0.12.0

# 数据下载（可选，用于Kaggle数据集）
# kagglehub>=0.1.0
# ===========================================
# LangChain和AI相关
# ===========================================

# LangChain核心
langchain>=0.3.0
langchain-core>=0.3.47
langchain-community==0.3.7
langchain-openai>=0.3.9
langchain-deepseek==0.1.3
langgraph==0.3.1

# MCP (Model Context Protocol) 支持
mcp>=1.0.0
fastmcp>=0.9.0
langchain-mcp-adapters>=0.1.0

# ===========================================
# 网络请求和API
# ===========================================

# HTTP客户端
httpx>=0.25.0
httpcore>=1.0.0
requests==2.31.0

# 豆包API (ByteDance)
volcengine-python-sdk>=1.0.0

# 数据库和API
psycopg2-binary==2.9.9

# ===========================================
# 异步支持
# ===========================================

aiohttp>=3.9.0
aiofiles>=23.0.0

# ===========================================
# 数据处理和工具
# ===========================================

# 数据处理
Pillow==10.1.0
openpyxl==3.1.2
PyYAML==6.0.1
xmltodict>=0.13.0

# 图像处理（天体图像分析）
opencv-python==4.8.1.78

# 数值计算优化
numba==0.58.1

# JSON处理
orjson>=3.10.1

# ===========================================
# 系统和工具
# ===========================================

# 系统工具
psutil==5.9.6
python-dotenv>=1.0.0
click==8.1.7

# 配置管理
pydantic>=2.5.2

# 时间处理
pytz==2023.3

# 包管理
packaging==23.2

# ===========================================
# 日志和监控
# ===========================================

loguru==0.7.2
memory-profiler==0.61.0
structlog>=23.0.0
rich>=13.0.0

# ===========================================
# 缓存和任务队列
# ===========================================

# 缓存
redis==5.0.1

# 任务队列
celery==5.3.4

# ===========================================
# 安全
# ===========================================

cryptography==41.0.7

# ===========================================
# 开发和测试
# ===========================================

# 测试框架
pytest==7.4.3
pytest-cov==4.1.0

# 代码质量
flake8==6.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
pylint==3.0.3

# 文档生成
sphinx==7.2.6

# ===========================================
# 模板和类型支持
# ===========================================

# 模板引擎
Jinja2==3.1.2

# 类型提示
typing-extensions>=4.8.0

# ===========================================
# 安装说明
# ===========================================
# 
# 安装命令：
# pip install -r requirements.txt
#
# 如果遇到版本冲突，可以尝试：
# pip install --upgrade pip
# pip install -r requirements.txt --force-reinstall
#
# 对于MCP功能，确保已正确配置环境变量：
# - TAVILY_API_KEY
# - LLM相关配置

# ===========================================
# FastAPI和Web服务
# ===========================================

# FastAPI框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# 认证和安全
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# 监控和指标
prometheus-client>=0.19.0

# 异步测试支持
pytest-asyncio>=0.21.0